Refactor of Deck of Cards
