+++ Gemfile
@@ -1,5 +1,3 @@
 source "https://rubygems.org"
 
 gem "rake"
-gem "capybara"
-gem "poltergeist"
